split data initialization into its own component
