--- components/SortByDate.js
@@ -1,3 +1,0 @@
-export default function sortByDate ()
-{
-}
